First Half of local TMP and read-only CFS mount speedup implementations

+++ bin/InitializeAltMTLsParallel.py
@@ -1,6 +1,4 @@
-#!/global/common/software/desi/cori/desiconda/20200801-1.4.0-spec/conda/bin/python -u
-from astropy.table import Table
-import astropy
+#!/global/common/software/desi/cori/desiconda/20211217-2.0.0/conda/bin/python -u
 import multiprocessing as mp
 from multiprocessing import Pool
 import logging
@@ -12,11 +10,16 @@ import numpy as np
 import os 
 from sys import argv
 from desiutil.log import get_logger
+import cProfile, pstats, io
+from pstats import SortKey
+profile = True
+pr = cProfile.Profile()
+
 log = get_logger()
 
-#List of healpixels for SV3 (dark and bright are same)
-
-print(argv)
+print('in python script REMOVE BEFORE PUSHING')
+log.info('in python script REMOVE BEFORE PUSHING')
+log.info(argv)
 seed = int(argv[1])
 ndir = int(argv[2])
 try:
@@ -26,6 +29,9 @@ except:
 obscon = argv[4]
 survey = argv[5]
 outputMTLDirBase = argv[6]
+if ('trunk' in outputMTLDirBase.lower()) or  ('ops' in outputMTLDirBase.lower()):
+    raise ValueError("In order to prevent accidental overwriting of the real MTLs, please remove \'ops\' and \'trunk\' from your MTL output directory")
+log.info('output directory for alternate MTLs: {0}'.format(outputMTLDirBase))
 HPListFile = argv[7]
 try:
     shuffleBrightPriorities = bool(int(argv[8]))
@@ -51,18 +57,31 @@ NNodes = int(argv[11])
 # If folder doesn't exist, then create it.
 if not os.path.isdir(outputMTLDirBase):
     os.makedirs(outputMTLDirBase)
+if os.path.exists(outputMTLDirBase + 'SeedFile'):
+    temp = open(outputMTLDirBase + 'SeedFile', 'r')
+    tempseed = int(temp.readlines()[0].split()[0])
+    log.info('Seed file already exists with seed {0:d}'.format(tempseed))
 
-with open(outputMTLDirBase + 'SeedFile', 'w') as f:
-    f.write(str(seed))
+    if int(tempseed) == int(seed):
+        log.info('random seeds are saved, continuing')
+    else:
+        raise RuntimeError('different random seed {0:d} provided than for initial generation {1:d}'.format(int(seed), int(tempseed)))
+else:
+    with open(outputMTLDirBase + 'SeedFile', 'w') as f:
+        f.write(str(seed))
 
 HPList = np.array(open(HPListFile,'r').readlines()[0].split(',')).astype(int)
-print(HPList)
+log.info('First healpixel: {0:d}'.format(HPList[0]))
+log.info('Last healpixel: {0:d}'.format(HPList[-1]))
+log.info('Number of healpixels: {0:d}'.format(int(len(HPList))))
+
 NodeID = int(os.getenv('SLURM_NODEID'))
 SlurmNProcs = int(os.getenv('SLURM_NPROCS'))
 
 NProc = int(NNodes*64)
 
-
+log.info('requested number of nodes: {0:d}'.format(NNodes))
+log.info('requested number of processes: {0:d}'.format(ndir))
 
 
 
@@ -73,22 +92,24 @@ outputMTLDir = outputMTLDirBase + "Univ{0:03d}/"
 HPList = np.array(open(HPListFile,'r').readlines()[0].split(',')).astype(int)
 print(HPList)
 
-# If folder doesn't exist, then create it.
-if not os.path.isdir(outputMTLDirBase):
-    os.makedirs(outputMTLDirBase)
-
-with open(outputMTLDirBase + 'SeedFile', 'w') as f:
-    f.write(str(seed))
 
 def procFunc(nproc):
-    print('starting fxn call')
+    log.info('starting fxn call')
+    if 'sv' in survey.lower():
+        log.info('sv survey')
+        mtlprestr = survey.lower()
+    else:
+        log.info('non sv survey')
+        mtlprestr = ''
+
+    if os.path.exists(outputMTLDir + '/{0}/{2}/{3}mtl-{2}-hp-{1}.ecsv'.format(survey.lower(),HPList[-1], obscon.lower(), mtlprestr)):
+        log.info('pathname')
+        log.info(outputMTLDir + '/{0}/{2}/{3}mtl-{2}-hp-{1}.ecsv'.format(survey.lower(),HPList[-1], obscon.lower(), mtlprestr))
+        return 42
+    log.info('still going')
     for hpnum in HPList:
-        if 'sv' in survey.lower():
-            mtlprestr = survey.lower()
-        else:
-            mtlprestr = ''
         exampleledger = exampleledgerbase + '/{0}/{2}/{3}mtl-{2}-hp-{1}.ecsv'.format(survey.lower(),hpnum, obscon.lower(), mtlprestr)
-        initializeAlternateMTLs(exampleledger, outputMTLDir, genSubset = nproc, seed = seed, obscon = obscon, survey = survey, saveBackup = True, hpnum = hpnum, overwrite = overwrite)
+        initializeAlternateMTLs(exampleledger, outputMTLDir, genSubset = nproc, seed = seed, obscon = obscon, survey = survey, saveBackup = True, hpnum = hpnum, overwrite = overwrite, profile = profile)
     print('ending function call')
     return 42           
 
@@ -114,15 +135,20 @@ assert(len(inds))
     
 print('b')
 print(inds)
+print('running on NProc = {0} processes'.format(NProc))
 p = Pool(NProc)
 atexit.register(p.close)
+log.info('running procFunc now on inds:')
+log.info(inds)
+pr.enable()
 result = p.map(procFunc,inds)
+pr.disable()
 print('c')
-
-
-
-
-
-
+s = io.StringIO()
+sortby = SortKey.CUMULATIVE
+ps = pstats.Stats(pr, stream=s).sort_stats(sortby)
+ps.print_stats()
+ps.dump_stats(outputMTLDirBase + '/InitializeAltMTLParallel.prof')
+print(s.getvalue())
